requirements.txt
selenium==3.141.0
pytest >= 6.2.2
allure-pytest>= 2.8.18
allure-python-commons>=2.8.18
pytest-rerunfailures>=9.1.1
pytest-logger>=0.5.1
Faker == 6.3.0
requests == 2.25.1
pytest-xdist == 2.2.1